Send magic links via new Notification Service
Rather than sending a message to an SQS service and having the separate
notification service send this email, we can just send it synchronously
as part of the request. The scale of our app and small userbase means
that the complexity of a distributed and super resilient service is not
worth the overhead that comes with it.