Send "application deadline reminder" email synchronously
Switch from sending messages to SQS for the notification service to pick
up, to sending emails synchronously when the "application deadline
reminder" task/job is run.

This is because we aren't at a scale where the technical complexity of a
queueing infrastructure does not give us enough benefit for the cost it
brings.