PROTO: in which Steven regrets not aligning with proto/base earlier
Thinks through what a well shaped auth flow might look like for apply
without additional microservices, databases and encryption.
- Redis -> Postgres
- Authenticator frontend -> Apply
- Account store -> Apply
- Fund store -> Apply
- Notification service -> Apply
- JWT -> Cookie

Note that it adds a proto magic link template, retrospecitively I should
have just lined up the vars it expects with the existing template but I
tweaked it to make it non-fund specific. The justification there is just
that you're authorising to access the "Apply for funding" service and
then you can apply for whatever fund it is you had the link for.

Spent a bit more time on this feature to try and demonstrate something
not completely trivial that we can demo how much less code and state a
developer needs to hold in their head to understand how it works.

Adds a few interactions like the application list page robustly handling
if there's an active round or not (again leaning on the database as much
as possible, reducing the surface area for business logic mistakes).
Proposed showing which user is logged in with a phase banner.

This also adds a starter template for assess showing how the apply tool
and the assess tool can both read from the same applications.

Merged at this stage to take advantage of and follow up on:
- using the same proto applications as the base branch
- making sure it aligns with the new service nav and upgraded govuk
  frontend
- assessment lists exploring if they can fit in the main "manage" domain
  area