Fix sqlalchemy/marshmallow warning on relationship
We've been getting this warning on app startup for a while now:

/app/.venv/lib/python3.10/site-packages/marshmallow_sqlalchemy/convert.py:124:
SAWarning: relationship 'AssessmentRecord.change_requests' will copy
column assessment_records.application_id to column
assessment_flag.application_id, which conflicts with relationship(s):
'AssessmentRecord.flags' (copies assessment_records.application_id to
assessment_flag.application_id). If this is not the intention, consider
if these relationships should be linked with back_populates, or if
viewonly=True should be applied to one or more if they are read-only.
For the less common case that foreign key constraints are partially
overlapping, the orm.foreign() annotation can be used to isolate the
columns that should be written towards.   To silence this warning, add
the parameter 'overlaps="flags"' to the
'AssessmentRecord.change_requests' relationship. (Background on this
warning at: https://sqlalche.me/e/20/qzyx) (This warning originated from
the configure_mappers() process, which was invoked automatically in
response to a user-initiated operation.)

This is because we added a (subset) relationship to AssessmentRecord. We
originally had a `flags` relationship to AssessmentFlags, and we added a
partial filtered relationship to the same table called
`change_requests`.

Having two relationships on the same FK mean that sqlalchemy doesn't
know which one to use when writing to the related table, which could
lead to unexpected behaviour.

By setting the `change_requests` relationship as read-only, we can
remove the warning.