Re-run migrations, add DataCollectionDefinition middleware
DataCollection rows used to have a reference to an (application) round.
If we want these to be generalised across any data collection (eg
post-award monitoring) that reference needs to go away, so have
restructured data collections slightly.

A definition of an application or reporting round will get a
DataCollectionDefinition, which has 0:x
DataCollectionDefinitionSections, which has 0:x
DataCollectionDefinitionQuestions.

An instance of an application or report will get a
DataCollectionInstance, which has 0:x DataCollectionInstanceSectionData
rows that contain a json blob of data for the relevant questions

I'd like to think this through all again if we build this for good -
this is mostly just in the spirit of moving fast rather than making
fully-thought-through decisions.