PROTO: rounds out comments for individual sections
Lifts some timeline styles from the hmrc frontend SCSS and fetches
comments when loading the application section page.

This concept feels like it could also be applied for scores and to the
application page (showing the total of scoring and commenting - although
for more complex assessments you might only want to show some of the
more recent behaviour).