Remove extraneous log message
We are getting two sentry errors for, seemingly, all errors in the app -
one with that actual exception and then another from this log message.

I think this is just extra noise - the sentry error from this log
message has no context, no useful info, and generates an extra alert.

Let's remove it and see how that affects sentry.