Add invoke task for sending application deadline reminders
This is currently handled by a lambda and writes to an SQS queue for
sending notifications:
https://github.com/communitiesuk/funding-service-design-workflows/blob/main/apps/pre-award/lambdas/application-deadline-reminder/application_reminder.py

We are retiring the notifications service and so rewriting the lambda as
a native part of this service so that it can talk to the db directly and
just send emails by talking to GOV.UK Notify directly.

Because I've touched a bunch of file I've kicked off some mypy
complaints, so also dealing with those here.

Sets up transactional test isolation for integration tests using the
pattern from:
https://docs.sqlalchemy.org/en/20/orm/session_transaction.html#joining-a-session-into-an-external-transaction-such-as-for-test-suites