Inline deployment pipeline
This patch starts by pulling in the deployment workflow job from our
shared workflows repository
(https://github.com/communitiesuk/funding-service-design-workflows)
before heavily modifying it to address a number of issues:

1) When deploying, the deployment and e2e tests should run as a single
batched unit, so that the tests are actually validating the deployed
changes. Before this patch, they would run with a break in the middle,
which meant if you had multiple pipelines the following could happen:

  * deploy 1
  * deploy 2
  * tests 1
  * tests 2

The tests for deploy 1 are then not really valid.

2) The pipeline didn't nicely support deploying copilot scheduled tasks
(jobs), which we now need to add.