Run db migrations as part of app deploy
Because db migrations and the deploy step run in a dependent chain
currently, you need to manually approve twice in the pipeline in order
to release to prod (because we have turned on this approval step).

By moving db migrations into the deploy job, we can reduce this to a
single approval.

It is _technically_ slightly problematic as it means that anything other
than the app deploy will be deployed before the db migrations have
finished.

But that is likely to only be an infrequent scheduled job in the
near/mid future, so the risk is low and I'm happy to take it.

Sorry if you're reading this commit because my assumption/assertion has
proved to be false.