PROTO: conditions from the question bank/ templates
Load conditions from the question bank so that we can prove they can be
specified on templates and then work for the actual data collection
definitions they're used to create.

This should also give us a really easy way to test this logic so can
probably supersede the invoke script to add custom questions.