Override prod variables for application deadline job
When deploying the main service, we override a bunch of environment
variables to account for the flask env name not matching the copilot env
name, and domains being different.

We didn't copy these across to the new scheduled job, and the mismatched
FLASK_ENV means that config isn't loading correctly in production.