PROTO-WIP: adds expression builder
Note that now we have the number data type its probably fine to remove
the `int()` type coercion we do in some of the expressions.

This is a WIP pass to show that there are validations and conditions,
there are some plans for adding some more human readable versions of
setting up basic configurations but they got messed up in other UI
changes - adding this as a starting point.