Ticket: https://mhclgdigital.atlassian.net/browse/FLS-1138
FLS-1138:

Running a scan using the Axe extension on Chrome revealed an accessibility issue in the apply tasklist. Specifically, a <p> tag was found as a child of a <dl> (description list) tag, which is not considered accessible. To resolve this issue, we have replaced the <p> tag with a <dt> tag.