FSPT-342 - FSPT-342 - Support basic auth in Form Runner
We wouldn't need to do anything if all requests to Form Runner originated from the user's browser, as the responsibility for credential entry would be passed to the user via a dialogue box. However, there is a backend request to Form Runner in Apply - specifically to get a rehydration token for session management. We need to make sure that we are passing basic auth credentials as part of this backend request. We implement this in line with existing precedents in Pre-Award and Post-Award - we hard-code the basic auth credentials in a URL we store in AWS Parameter Store (e.g., http://username:password@example.com) and then import that variable as a secret that can be leveraged with no change required to application code.