Bugfix - Fix failing Apply E2E tests
These tests have been failing since basic auth was added to Form Runner. This is because the continue_application function, which redirects users to the Form Runner, is core to Apply.

By switching FORMS_SERVICE_PUBLIC_HOST from a Copilot variable to a secret, we can store a basic auth-embedded version of the Form Runner URL in the AWS Parameter Store, meaning we can automate our way through the authentication barrier. As a byproduct, Apply users in pre-prod environments will also be spared the pain of basic auth. As the main point of basic auth is to block web crawlers, this seems okay.