FLS-1424:
While investigating why the columns were not appearing in the Applicant Response section of Assess, I discovered that the issue stemmed from the get_data method. This method was used to call an endpoint (/sub_criteria_themes/{application_id}) that fetched the JSON subcriteria. Although it was only intended to retrieve the data, it implicitly ordered the fields based on the keys (e.g., "kzCRNQ"), which affected the column rendering.

Additionally, since we've transitioned to a monolith infrastructure, it's now best practice to avoid using endpoints to interact with other parts of Pre-Award (such as assessment-store and application-store). Instead, we should directly query the relevant services or use their existing methods. In line with this, I replaced the endpoint call with a direct invocation of the appropriate route method within the service get_application_json_and_sub_criterias.

- Replaced call to endpoint (/sub_criteria_themes/{application_id}) with call to the method itself get_application_json_and_sub_criterias
- Removed get_all_sub_criterias_with_application_json from data_services since it's only calling the endpoint, instead replaced where the removed method (get_all_sub_criterias_with_application_json) was being called with call to the intended route's method get_application_json_and_sub_criterias
- Removed the endpoint /sub_criteria_themes/{application_id} from the api.yml and from default.py since it's not being used anywhere.