Use AssessmentRecord not Applications model when retrieving all comments
Retrieving comments happens in the assessment store scope. Using the Applications model from application store breaks that separation. We don't need to do that as we have AssessmentRecord in scope, with application reference, fund and round ID, so let's use that instead. While longer-term consolidation is appropriate, for now Pre-Award code is divided up according to the extinct microservices, and we should follow this pattern here.