Fixes issue whereby assessment comments cannot be saved or retrieved for multilingual funds e.g., COF
This issue originated from a recent refactor that moved the logic for retrieving comments from application store to assessment store. The refactor included a call to an existing method get_all_subcriteria which expects language as a string e.g., 'en'; the new code passed it as an enum object. For funds and rounds without Welsh configured as a language option, we simply ignored the language variable inside get_all_subcriteria and all was well. This meant that fortunately we had no issues in production, as only COF supports Welsh, and there were no COF rounds active during the time the bug was live in production.

This change ensures that we pass language as a string to get_all_subcriteria, ensuring the function can execute without error even in the case of multilingual funds like COF.