Testing removal of UAT sidecar (not to be merged)
We are seeing some 502 Gateway error in pre-prod environments. Removing certain roles from the database seems to fix this. We are not sure why but we suspect the nginx sidecar we are using may be causing problems, as it is deprecated (not updated in six years) and we have seen issues with basic auth failing if response cookie is too large. To test this theory, we are going to re-add the roles we removed from the UAT database to fix the issue, back into the UAT database, deploy a sidecar-less UAT service and see if we get the 502 still. If not, then we can feel more confident about deploying to prod and unblocking the pipeline while we figure out a longer-term fix for the sidecar issue.