Bugfix - Allowing multiple rounds from multiple config files for the same fund to be loaded via FAB fund round loader script
This functionality was lost during recent refactoring work that was done to support a new API endpoint. The logic is all still a bit janky, with the possibility of specifying rounds both within a single config file and across multiple config files, but this fix at least reverts us to the functionality we enjoyed before.