FLS-1455 - Script to migrate JSON forms from fsd_config directory into Pre-Award database
This script should be run locally. You need to pass the connection string of the database you want to connect to e.g., if local database then 'postgresql://postgres:password@localhost:5432/pre_award_stores'. To run against a deployed database, authenticate with AWS (using e.g., 'aws-vault exec test'), create an SSH-style tunnel using 'funding-service-design-workflows/scripts/aws_connect_db.sh' to expose the database on localhost, and then take the connection string that script prints. You'll need to change the prefix 'postgres' for 'postgresql' for this migration script to run. You'll also need the digital-form-builder-adapter repo installed locally as a sister directory to this repo, as that's where we get the form JSONs from, and that's the reason we run this script locally.