FLS-1450 - Remove JSON from GET /forms endpoint
Previously we returned a list of forms along with all of their attributes from the GET /forms endpoint, including draft_json and published_json, which can be very large. The endpoint is only hit by the Form Designer and the draft_json and published_json attributes aren't accessed on the requester side, plus the amount of data being sent over the network meant the API call took an age to complete. This change removes those attributes from the response.