FLS-1453 - Add display_name to form_definition and rename name to url_path
Currently we 'name' forms in two places: at the point they are created in Form Designer, where we have to enter a 'Title' value; and at the point they are uploaded as a template in FAB, where we enter both a 'Template name' (how the form is referred to in FAB) and a 'Task name' (how the form is displayed in Apply).

Our forms re-architecture work has to this point omitted the capability of naming a form beyond its URL path. This PR addresses that, ensuring that form names can be rendered across different services without having to reverse engineer them from URL paths, by introducing a specific new field 'display_name' to the 'form_definition' table. The field 'name' is renamed to 'url_path' to disambiguate it.