FLS-1449 - Pre-Award API - adaptations in response to consuming service requirements
It is okay to modify the API at this point as it is not being consumed in production. The initial integration work has been done for each of the consuming services - Form Designer, Form Runner and FAB - and it has become clear that some changes are required in the API to avoid having to do hacky workarounds in those services. For example, we often want to get the display name of a particular form, and with the original API design that means getting all forms and finding the one with the matching ID, and taking the display name there, because the get draft form and get published form endpoints simply return JSON form configurations and no form metadata.

This commit is focused on rectifying that issue, in both the 'get draft form' and 'get published form' endpoints. Form Designer requires draft form configurations alongside form display names, while FAB requires published form configurations alongside form display names.