FLS-1456 - Override Form Designer host in Production Authenticator to fix post sign-in redirect
After successfully signing in with Microsoft, Authenticator redirects users back to the app they came from. The exact return URL is specified in the AWS Copilot manifest file, and in the case of Form Designer it is currently https://form-designer.access-funding.${COPILOT_ENVIRONMENT_NAME}.communities.gov.uk across all environments. This does not work in Prod because the interpolation gives https://form-designer.access-funding.prod.communities.gov.uk which is not correct - it should be https://form-designer.access-funding.communities.gov.uk. Thus we should override the Form Designer return URL in the Prod configuration to fix this.