FLS-1513 - Upscale resources in Test environment to match UAT
E2E tests are constantly failing in Test environment, but passing in UAT. The obvious culprit seems to be the difference in resource allocation between the environments. Upscaling resources does not tackle the root cause, but as failing E2E tests are blocking critical deployment pipelines, it is potentially a quick and pragmatic solution. We will similarly upscale Form Runner in the Test environment.