BAU - Re-open DPIF R4 for one full working day as per Tom Lowndes
We have had a request from the grant team to re-open DPIF R4 to allow two LAs to submit their applications. This commit changes the round opening dates for DPIF R4 to reopen the round between 9am and 5pm on 28th November. We ensure application reminder emails will not be erroneously sent by updating the application_reminder_sent value to True in the application config. We will need to manually run a script in production to enact this change once hte code is deployed.