FLS-1636 - Make CTDF fund competitive (currently uncompeted)
We are creating a new dummy fund round CTDF CR2 and the idea was that it would be competitive by default as no uncompeted fund features are being tested in the E2E tests. Although we set the funding_type as COMPETITIVE in the new fund round config, the funding_type UNCOMPETED from CTDF CR1 overrode that owing to the way fund rounds are loaded, and also the fact that each fund can only have funding_type - really the entire idea of having multiple round configs for a single fund, each containing their own version of the fund config, doesn't make a lot of sense, but that's a pain point for another day.